+++ back/src/main/java/fr/centralesupelec/thuv/model/Course.java
@@ -139,4 +139,6 @@ public class Course {
     private LocalDateTime externalAccessExpirationDate;
     
     private String uid;
+    @Column(nullable = false)
+    private Long numberOfUsers = 0L;
 }